extract RefreshMessageTypes from RefreshWorker
to not trigger execution of worker code when only RefreshMessageTypes are imported in another file
